pyproject.toml
# StackSentry — Security Audit Framework
# Project metadata and tool configuration
 
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
 
[project]
name = "stacksentry"
version = "1.0.0"
description = "Automated web application security configuration assessment and remediation"
authors = [{name = "Victor Chukwuemeka Onwuegbuchulem", email = "vio0264@my.londonmet.ac.uk"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
 
dependencies = [
    "requests==2.31.0",
    "urllib3>=1.21.1,<2.1",
    "docker==7.0.0",
    "paramiko==3.4.0",
    "PyYAML==6.0.1",
    "reportlab==4.2.2",
    "anthropic>=0.50.0",
    "python-dotenv>=1.0.0",
]
 
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]
 
[project.scripts]
stacksentry = "sec_audit.cli:main"
 
[tool.hatch.build.targets.wheel]
packages = ["sec_audit", "checks", "scanners", "reporting", "storage", "remediation"]
 
# ── Pytest configuration ──────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::cryptography.utils.CryptographyDeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
markers = [
    "slow: marks tests as slow (deselect with -m not slow)",
    "integration: marks tests that require real Docker or SSH targets",
]
 
# ── Coverage configuration ────────────────────────────────────────────────────
[tool.coverage.run]
source = ["checks", "sec_audit", "scanners", "reporting"]
omit = ["tests/*", "*/migrations/*"]
 
[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 70